#244 Remove MolecularType and Use MoleculePool
