+++ ecell4/core/LatticeSpaceVectorImpl.cpp
@@ -1,5 +1,5 @@
 #include "Context.hpp"
-#include "MolecularType.hpp"
+#include "MoleculePool.hpp"
 #include "VacantType.hpp"
 #include "StructureType.hpp"
 #include "InterfaceType.hpp"
@@ -17,9 +17,9 @@ LatticeSpaceVectorImpl::LatticeSpaceVectorImpl(
     std::stringstream ss;
     ss << voxel_radius_;
     border_ = boost::shared_ptr<VoxelPool>(
-            new MolecularType(Species("Border", ss.str(), "0"), vacant_));
+            new MoleculePool(Species("Border", ss.str(), "0"), vacant_));
     periodic_ = boost::shared_ptr<VoxelPool>(
-            new MolecularType(Species("Periodic", ss.str(), "0"), vacant_));
+            new MoleculePool(Species("Periodic", ss.str(), "0"), vacant_));
 
     initialize_voxels(is_periodic_);
 }
